requirements.txt
Jinja2
black
click
isort
pelican
pelican-alias
python-slugify
tweepy
